Feat: 가입 신청 조회 Dto 항목 추가 및 다중 수락, 거절 로직 추가

+++ src/main/java/GDG/whatssue/domain/clubjoinrequest/service/ClubJoinManageService.java
@@ -1,12 +1,16 @@
 package GDG.whatssue.domain.clubjoinrequest.service;
 
+import GDG.whatssue.domain.club.exception.ClubErrorCode;
 import GDG.whatssue.domain.clubjoinrequest.entity.ClubJoinRequest;
 import GDG.whatssue.domain.member.entity.ClubMember;
 import GDG.whatssue.domain.clubjoinrequest.repository.ClubJoinRequestRepository;
+import GDG.whatssue.domain.member.exception.ClubMemberErrorCode;
 import GDG.whatssue.domain.member.repository.ClubMemberRepository;
 import GDG.whatssue.domain.clubjoinrequest.dto.ClubJoinRequestGetDto;
+import GDG.whatssue.global.error.CommonException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.tool.schema.spi.CommandAcceptanceException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,24 +24,41 @@ public class ClubJoinManageService {
     private final ClubMemberRepository clubMemberRepository;
 
     @Transactional
-    public void acceptResponse(Long clubJoinRequestId) {//수락
+    public void acceptResponse(List<Long> clubJoinRequestsId,Long clubId) {//수락
         // 클럽 가입 요청 엔티티 조회
-        ClubJoinRequest clubJoinRequest = clubJoinRequestRepository.findById(clubJoinRequestId)
-                .orElseThrow(() -> new IllegalArgumentException("클럽 가입 요청을 찾을 수 없습니다: " + clubJoinRequestId));
+        for (Long clubJoinRequestId : clubJoinRequestsId) {
+            ClubJoinRequest clubJoinRequest = clubJoinRequestRepository.findById(clubJoinRequestId)
+                    .orElseThrow(() -> new CommonException(ClubErrorCode.EX3102)); //존재하지 않는 클럽 가입 신청
 
-        ClubMember member = ClubMember.newMember(clubJoinRequest.getClub(), clubJoinRequest.getUser());
+            if(clubJoinRequest.getClub().getId()!=clubId){
+                throw new CommonException(ClubErrorCode.EX3207); // 다른 클럽의 가입 신청에 접근 불가능
+            }
 
-        clubMemberRepository.save(member);
-        clubJoinRequestRepository.delete(clubJoinRequest);
+            if(clubMemberRepository.findByClub_IdAndUser_UserId(clubId,clubJoinRequest.getUser().getUserId()).isPresent()) {
+                throw new CommonException(ClubErrorCode.EX3200); // 이미 가입한 모임
+            }
+            ClubMember member = ClubMember.newMember(clubJoinRequest.getClub(), clubJoinRequest.getUser());
+
+            clubMemberRepository.save(member);
+            clubJoinRequestRepository.delete(clubJoinRequest);
+        }
     }
-
     @Transactional
-    public void denyResponse(Long clubJoinRequestId) {//거절
+    public void denyResponse(List<Long> clubJoinRequestsId, Long clubId) {//거절
         // 클럽 가입 요청 엔티티 조회
-        ClubJoinRequest clubJoinRequest = clubJoinRequestRepository.findById(clubJoinRequestId)
-                .orElseThrow(() -> new IllegalArgumentException("클럽 가입 요청을 찾을 수 없습니다: " + clubJoinRequestId));
+        for (Long clubJoinRequestId : clubJoinRequestsId) {
+            ClubJoinRequest clubJoinRequest = clubJoinRequestRepository.findById(clubJoinRequestId)
+                    .orElseThrow(() -> new CommonException(ClubErrorCode.EX3102)); //존재하지 않는 클럽 가입 신청
 
-        clubJoinRequestRepository.delete(clubJoinRequest);
+            if(clubJoinRequest.getClub().getId()!=clubId){
+                throw new CommonException(ClubErrorCode.EX3207); // 다른 클럽의 가입 신청에 접근 불가능
+            }
+
+            if(clubMemberRepository.findByClub_IdAndUser_UserId(clubId,clubJoinRequest.getUser().getUserId()).isPresent()) {
+                throw new CommonException(ClubErrorCode.EX3200); // 이미 가입한 모임
+            }
+            clubJoinRequestRepository.delete(clubJoinRequest);
+        }
     }
 
     @Transactional
@@ -56,8 +77,10 @@ public class ClubJoinManageService {
     private ClubJoinRequestGetDto mapToDto(ClubJoinRequest clubJoinRequest) {
         return ClubJoinRequestGetDto.builder()
                 .id(clubJoinRequest.getId())
-                .club(clubJoinRequest.getClub())
-                .user(clubJoinRequest.getUser())
+                .clubId(clubJoinRequest.getClub().getId())
+                .userId(clubJoinRequest.getUser().getUserId())
+                .userName(clubJoinRequest.getUser().getUserName())
+                .createdAt(clubJoinRequest.getCreateAt())
                 .build();
     }
